+++ src/components/pages/LandingPage/Deals/_deals.scss
