+++ src/components/pages/LandingPage/NewArrivals/_newArrivals.scss
