+++ src/components/pages/LandingPage/RecentlyBought/ProductCard/_productCard.scss
