+++ src/components/pages/LandingPage/RecentlyBought/_recentlyBought.scss
