+++ src/components/pages/LandingPage/Trending/TrendingCard/_trendingCard.scss
