+++ src/components/pages/LandingPage/Trending/_trending.scss
